signout and products & orders components been added

+++ front-end/admin-app/src/actions/auth_action.js
@@ -1,31 +1,31 @@
-import { authConstants } from "./constants";
-import Axios from "../helpers/axios";
+import { authConstants } from './constants'
+import Axios from '../helpers/axios'
 
 export const login = (user) => {
-  console.log(user);
+  console.log(user)
   return async (dispatch) => {
-    dispatch({ type: authConstants.LOGIN_REQUEST });
-    const res = await Axios.post("/admin/login", {
+    dispatch({ type: authConstants.LOGIN_REQUEST })
+    const res = await Axios.post('/admin/signin', {
       ...user,
-    });
+    })
 
     if (res.status === 200) {
-      const { token, user } = res.data;
-      localStorage.setItem("token", token);
-      localStorage.setItem("user", JSON.stringify(user));
+      const { token, user } = res.data
+      localStorage.setItem('token', token)
+      localStorage.setItem('user', JSON.stringify(user))
       dispatch({
         type: authConstants.LOGIN_SUCCESS,
         payload: {
           token,
-          user
+          user,
         },
-      });
+      })
     } else {
       if (res.status === 400) {
         dispatch({
           type: authConstants.LOGIN_FAILURE,
           payload: { error: res.data.error },
-        });
+        })
       }
     }
 
@@ -34,39 +34,49 @@ export const login = (user) => {
       payload: {
         ...user,
       },
-    });
-  };
-};
-
-
+    })
+  }
+}
 
 export const isUserLoggedIn = () => {
-  return async dispatch => {
-    const token = localStorage.getItem("token");
+  return async (dispatch) => {
+    const token = localStorage.getItem('token')
 
     if (token) {
-      const user = JSON.parse(localStorage.getItem("user")) ;
+      const user = JSON.parse(localStorage.getItem('user'))
       dispatch({
         type: authConstants.LOGIN_SUCCESS,
         payload: {
           token,
           user,
         },
-      });
+      })
     } else {
       dispatch({
         type: authConstants.LOGIN_FAILURE,
-        payload: { error: "fail to login" },
-      });
+        payload: { error: 'fail to login' },
+      })
     }
-  };
-};
+  }
+}
 
 export const signout = () => {
-    return async dispatch => {
-        localStorage.clear();
-        dispatch({
-            type: authConstants.LOGOUT_REQUEST
-        })
+  return async (dispatch) => {
+    dispatch({ type: authConstants.LOGOUT_REQUEST })
+    const res = await Axios.post('/admin/signout')
+
+    if (res.status === 200) {
+      localStorage.clear()
+      dispatch({
+        type: authConstants.LOGOUT_SUCCESS,
+      })
+    } else {
+      dispatch({
+        type: authConstants.LOGOUT_FAILURE,
+        payload: {
+          error: res.data.error,
+        },
+      })
     }
+  }
 }
